+++ example-functional/src/main/java/com/dan323/functional/data/optional/Maybe.java
@@ -1,11 +1,7 @@
 package com.dan323.functional.data.optional;
 
-import com.dan323.functional.annotation.Functor;
-import com.dan323.functional.annotation.funcs.IFunctor;
-
 import java.util.function.Function;
 
-@Functor
 public sealed interface Maybe<A> permits Nothing, Just {
 
     <C> C maybe(Function<A, C> f, C constant);