+++ example-functional/src/main/java/com/dan323/functional/data/writer/LoggerMonad.java
@@ -1,8 +1,5 @@
 package com.dan323.functional.data.writer;
 
-import com.dan323.functional.annotation.Monad;
-
-@Monad
 public final class LoggerMonad extends WriterMonad<String> {
 
     public LoggerMonad() {