+++ functional-definitions/functional-compiler/src/test/java/com/dan323/functional/annotation/compiler/traversal/TraversalMock.java
@@ -1,0 +1,23 @@
+package com.dan323.functional.annotation.compiler.traversal;
+
+import com.dan323.functional.annotation.Traversal;
+import com.dan323.functional.annotation.compiler.util.ApplicativeUtil;
+import com.dan323.functional.annotation.funcs.IApplicative;
+import com.dan323.functional.annotation.funcs.ITraversal;
+
+import java.util.Optional;
+import java.util.function.Function;
+
+@Traversal
+public class TraversalMock implements ITraversal<Optional<?>> {
+
+    public <K, A> K traverse(IApplicative<K> applicative, Function<A, K> fun, Optional<A> elem) {
+        return elem.map(fun).orElse(ApplicativeUtil.pure(applicative, Optional.empty()));
+    }
+
+    @Override
+    public Class<Optional<?>> getClassAtRuntime() {
+        return (Class<Optional<?>>) (Class<?>) Optional.class;
+    }
+}
+